Refactored analytics to its own component
