+++ src/app/analytics-feature/analytics-feature.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AnalyticsComponent } from './analytics/analytics.component';
+import { AnalyticsService } from './analytics/analytics.service';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [
+    AnalyticsComponent
+  ],
+  providers: [
+    AnalyticsService
+  ]
+})
+export class AnalyticsFeatureModule { }
