+++ src/app/analytics-feature/analytics/analytics.component.html
@@ -85,6 +85,8 @@
 
       </div>
       <!--End Grid row cards-->
+
+      <h4>Session Details</h4>
       <hr>
       <!--Start Grid row chart-->
       <div class="row">
