+++ src/app/analytics-feature/analytics/analytics.component.scss
