+++ src/app/analytics-feature/model/analytics.model.ts
