+++ src/app/analytics-feature/model/stats.model.ts
