+++ src/app/model/model.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { Account } from "./account.model";
-import { Analytics } from "./analytics.model";
+import { Analytics } from "../analytics-feature/model/analytics.model";
 
 @Injectable()
 export class ModelService {
